awx, eda, auth and common custom commands
