+++ cypress/support/commands.ts
@@ -3,7 +3,6 @@
 /// <reference types="cypress" />
 import '@cypress/code-coverage/support';
 import { SetOptional } from 'type-fest';
-import { randomString } from '../../framework/utils/random-string';
 import { Inventory } from '../../frontend/awx/interfaces/Inventory';
 import { Organization } from '../../frontend/awx/interfaces/Organization';
 import { Project } from '../../frontend/awx/interfaces/Project';
@@ -18,12 +17,14 @@ import {
 import { EdaCredential } from '../../frontend/eda/interfaces/EdaCredential';
 import { EdaDecisionEnvironment } from '../../frontend/eda/interfaces/EdaDecisionEnvironment';
 import { EdaProject } from '../../frontend/eda/interfaces/EdaProject';
-import { EdaResult } from '../../frontend/eda/interfaces/EdaResult';
 import { EdaRulebook } from '../../frontend/eda/interfaces/EdaRulebook';
 import { EdaRulebookActivation } from '../../frontend/eda/interfaces/EdaRulebookActivation';
 import { EdaUser } from '../../frontend/eda/interfaces/EdaUser';
-import './rest-commands';
 import { Label } from '../../frontend/awx/interfaces/Label';
+import './rest-commands';
+import './auth';
+import './eda-commands';
+import './awx-commands';
 
 declare global {
   namespace Cypress {
@@ -300,775 +301,3 @@ declare global {
     }
   }
 }
-
-Cypress.Commands.add(
-  'login',
-  (server: string, username: string, password: string, serverType: string) => {
-    window.localStorage.setItem('theme', 'light');
-    window.localStorage.setItem('disclaimer', 'true');
-
-    if (serverType === 'EDA server' && Cypress.env('TEST_STANDALONE') === true) {
-      // Standalone EDA login
-      cy.visit(`/login`, {
-        retryOnStatusCodeFailure: true,
-        retryOnNetworkFailure: true,
-      });
-      cy.typeInputByLabel(/^Username$/, username);
-      cy.typeInputByLabel(/^Password$/, password);
-      cy.get('button[type=submit]').click();
-      return;
-    }
-
-    cy.visit(`/automation-servers`, {
-      retryOnStatusCodeFailure: true,
-      retryOnNetworkFailure: true,
-    });
-
-    cy.clickButton(/^Add automation server$/);
-    cy.typeInputByLabel(/^Name$/, 'E2E');
-    cy.typeInputByLabel(/^Url$/, server);
-    cy.get('.pf-c-select__toggle').click();
-    cy.clickButton(serverType);
-    cy.get('button[type=submit]').click();
-
-    cy.contains('a', /^E2E$/).click();
-    cy.typeInputByLabel(/^Username$/, username);
-    cy.typeInputByLabel(/^Password$/, password);
-    cy.get('button[type=submit]').click();
-  }
-);
-
-Cypress.Commands.add('edaLogout', () => {
-  cy.get('.pf-c-dropdown__toggle')
-    .eq(1)
-    .click()
-    .then(() => {
-      cy.get('ul>li>a').contains('Logout').click();
-    });
-});
-
-Cypress.Commands.add('awxLogin', () => {
-  cy.session(
-    'AWX',
-    () => {
-      cy.login(
-        Cypress.env('AWX_SERVER') as string,
-        Cypress.env('AWX_USERNAME') as string,
-        Cypress.env('AWX_PASSWORD') as string,
-        'AWX Ansible server'
-      );
-      cy.hasTitle(/^Welcome to AWX$/);
-    },
-    {
-      cacheAcrossSpecs: true,
-      validate: () => {
-        cy.request({ method: 'GET', url: '/api/v2/me/' });
-      },
-    }
-  );
-  cy.visit(`/ui_next`, { retryOnStatusCodeFailure: true, retryOnNetworkFailure: true });
-});
-
-Cypress.Commands.add('edaLogin', () => {
-  cy.session(
-    'EDA',
-    () => {
-      cy.login(
-        Cypress.env('EDA_SERVER') as string,
-        Cypress.env('EDA_USERNAME') as string,
-        Cypress.env('EDA_PASSWORD') as string,
-        'EDA server'
-      );
-      cy.hasTitle(/^Welcome to EDA Server$/);
-    },
-    {
-      cacheAcrossSpecs: true,
-      validate: () => {
-        cy.request({ method: 'GET', url: '/api/eda/v1/users/me/' });
-      },
-    }
-  );
-  cy.visit(`/eda`, { retryOnStatusCodeFailure: true, retryOnNetworkFailure: true });
-});
-
-Cypress.Commands.add('getFormGroupByLabel', (label: string | RegExp) => {
-  cy.contains('.pf-c-form__label-text', label).parent().parent().parent();
-});
-
-Cypress.Commands.add('getInputByLabel', (label: string | RegExp) => {
-  cy.contains('.pf-c-form__label-text', label)
-    .parent()
-    .invoke('attr', 'for')
-    .then((id: string | undefined) => {
-      if (id) {
-        cy.get('#' + id).should('be.enabled');
-      }
-    });
-});
-
-Cypress.Commands.add('getCheckboxByLabel', (label: string | RegExp) => {
-  cy.contains('.pf-c-check__label', label)
-    .invoke('attr', 'for')
-    .then((id: string | undefined) => {
-      if (id) {
-        cy.get('#' + id);
-      }
-    });
-});
-
-Cypress.Commands.add('typeInputByLabel', (label: string | RegExp, text: string) => {
-  cy.getInputByLabel(label).clear().type(text, { delay: 0 });
-});
-
-Cypress.Commands.add('selectDropdownOptionByLabel', (label: string | RegExp, text: string) => {
-  cy.getFormGroupByLabel(label).within(() => {
-    // Click button once it is enabled. Async loading of select will make it disabled until loaded.
-    cy.get('button[aria-label="Options menu"]').should('be.enabled').click();
-
-    // If the select menu contains a serach, then search for the text
-    cy.get('.pf-c-select__menu').then((selectMenu) => {
-      if (selectMenu.find('.pf-m-search').length > 0) {
-        cy.get('.pf-m-search').clear().type(text, { delay: 0 });
-      }
-    });
-
-    cy.get('.pf-c-select__menu').within(() => {
-      cy.contains('button', text).click();
-    });
-  });
-});
-
-Cypress.Commands.add('selectToolbarFilterType', (text: string | RegExp) => {
-  cy.get('#filter-form-group').within(() => {
-    cy.get('.pf-c-select').should('not.be.disabled').click();
-    cy.get('.pf-c-select__menu').within(() => {
-      cy.clickButton(text);
-    });
-  });
-});
-
-Cypress.Commands.add('filterTableByText', (text: string) => {
-  cy.get('#filter-input').within(() => {
-    cy.get('input').clear().type(text, { delay: 0 });
-  });
-  cy.get('[aria-label="apply filter"]').click();
-});
-
-Cypress.Commands.add('filterTableByTypeAndText', (filterLabel: string | RegExp, text: string) => {
-  cy.selectToolbarFilterType(filterLabel);
-  cy.filterTableByText(text);
-});
-
-Cypress.Commands.add('clickLink', (label: string | RegExp) => {
-  cy.contains('a:not(:disabled):not(:hidden)', label).click();
-});
-
-Cypress.Commands.add('clickTab', (label: string | RegExp) => {
-  cy.contains('button[role="tab"]', label).click();
-});
-
-Cypress.Commands.add('clickButton', (label: string | RegExp) => {
-  cy.contains('button:not(:disabled):not(:hidden)', label).click();
-});
-
-Cypress.Commands.add('navigateTo', (label: string | RegExp) => {
-  cy.get('#page-sidebar').then((c) => {
-    if (c.hasClass('pf-m-collapsed')) {
-      cy.get('#nav-toggle').click();
-    }
-  });
-  cy.contains('.pf-c-nav__link', label).click();
-  cy.get('#page-sidebar').then((c) => {
-    if (!c.hasClass('pf-m-collapsed')) {
-      cy.get('#nav-toggle').click();
-    }
-  });
-  cy.get('#refresh').click();
-});
-
-Cypress.Commands.add('hasTitle', (label: string | RegExp) => {
-  cy.contains('.pf-c-title', label);
-});
-
-Cypress.Commands.add('hasAlert', (label: string | RegExp) => {
-  cy.contains('.pf-c-alert__title', label);
-});
-
-Cypress.Commands.add('hasTooltip', (label: string | RegExp) => {
-  cy.contains('.pf-c-tooltip__content', label);
-});
-
-Cypress.Commands.add('clickToolbarKebabAction', (label: string | RegExp) => {
-  cy.get('.page-table-toolbar').within(() => {
-    cy.get('.toggle-kebab').click().get('.pf-c-dropdown__menu-item').contains(label).click();
-  });
-});
-
-Cypress.Commands.add('clickTableRow', (name: string | RegExp, filter?: boolean) => {
-  if (filter !== false && typeof name === 'string') {
-    cy.filterTableByText(name);
-  }
-  cy.contains('td', name).within(() => {
-    cy.get('a').click();
-  });
-});
-
-Cypress.Commands.add('getTableRowByText', (name: string | RegExp, filter?: boolean) => {
-  if (filter !== false && typeof name === 'string') {
-    cy.filterTableByText(name);
-  }
-  cy.contains('tr', name);
-});
-
-Cypress.Commands.add(
-  'clickTableRowKebabAction',
-  (name: string | RegExp, label: string | RegExp, filter?: boolean) => {
-    cy.getTableRowByText(name, filter).within(() => {
-      cy.get('.pf-c-dropdown__toggle').click();
-      cy.contains('.pf-c-dropdown__menu-item', label)
-        .should('not.be.disabled')
-        .should('not.have.attr', 'aria-disabled', 'true')
-        .click();
-    });
-  }
-);
-
-Cypress.Commands.add(
-  'clickTableRowPinnedAction',
-  (name: string | RegExp, label: string, filter?: boolean) => {
-    cy.getTableRowByText(name, filter).within(() => {
-      cy.get(`#${label.toLowerCase().split(' ').join('-')}`)
-        .should('not.be.disabled')
-        .should('not.have.attr', 'aria-disabled', 'true')
-        .click();
-    });
-  }
-);
-
-Cypress.Commands.add(
-  'clickTableRowActionIcon',
-  (name: string | RegExp, ariaLabel: string, filter?: boolean) => {
-    cy.getTableRowByText(name, filter).within(() => {
-      cy.get(`button[aria-label="${ariaLabel}"]`)
-        .should('not.be.disabled')
-        .should('not.have.attr', 'aria-disabled', 'true')
-        .click();
-    });
-  }
-);
-
-Cypress.Commands.add('tableHasRowWithSuccess', (name: string | RegExp, filter?: boolean) => {
-  cy.getTableRowByText(name, filter).within(() => {
-    cy.get('[data-label="Status"]').should('contain', 'Successful');
-  });
-});
-
-Cypress.Commands.add('selectTableRow', (name: string | RegExp, filter?: boolean) => {
-  cy.getTableRowByText(name, filter).within(() => {
-    cy.get('input[type=checkbox]').click();
-  });
-});
-
-Cypress.Commands.add('getDialog', () => {
-  cy.get('div[data-ouia-component-type="PF4/ModalContent"]');
-});
-
-Cypress.Commands.add('selectTableRowInDialog', (name: string | RegExp, filter?: boolean) => {
-  cy.getDialog().within(() => {
-    cy.getTableRowByText(name, filter).within(() => {
-      cy.get('input[type=checkbox]').click();
-    });
-  });
-});
-
-Cypress.Commands.add('expandTableRow', (name: string | RegExp, filter?: boolean) => {
-  cy.getTableRowByText(name, filter).within(() => {
-    cy.get('button[id^="expand-toggle"]').click();
-  });
-});
-
-Cypress.Commands.add(
-  'hasDetail',
-  (detailTerm: string | RegExp, detailDescription: string | RegExp) => {
-    cy.contains('dt', detailTerm).next().should('contain', detailDescription);
-  }
-);
-
-Cypress.Commands.add('clickModalButton', (label: string | RegExp) => {
-  cy.getDialog().within(() => {
-    cy.contains('button', label).click();
-  });
-});
-
-Cypress.Commands.add('clickModalConfirmCheckbox', () => {
-  cy.getDialog().within(() => {
-    cy.get('input[id="confirm"]').click();
-  });
-});
-
-Cypress.Commands.add('assertModalSuccess', () => {
-  cy.getDialog().within(() => {
-    cy.get('tbody>tr')
-      .find('[data-label="Status"]')
-      .each(($li) => {
-        cy.wrap($li).should('contain', 'Success');
-      });
-  });
-});
-
-Cypress.Commands.add('clickPageAction', (label: string | RegExp) => {
-  cy.get('.toggle-kebab').click().get('.pf-c-dropdown__menu-item').contains(label).click();
-});
-
-/**Resources for testing AWX*/
-Cypress.Commands.add('createAwxOrganization', () => {
-  cy.requestPost<Organization>('/api/v2/organizations/', {
-    name: 'E2E Organization ' + randomString(4),
-  }).then((organization) => organization);
-});
-
-Cypress.Commands.add('deleteAwxOrganization', (organization: Organization) => {
-  cy.requestDelete(`/api/v2/organizations/${organization.id}/`, true);
-});
-
-Cypress.Commands.add('createAwxTeam', (organization: Organization) => {
-  cy.requestPost<Team>('/api/v2/teams/', {
-    name: 'E2E Team ' + randomString(4),
-    organization: organization.id,
-  }).then((team) => team);
-});
-
-Cypress.Commands.add('deleteAwxTeam', (team: Team) => {
-  if (team.id) {
-    cy.requestDelete(`/api/v2/teams/${team.id.toString()}/`, true);
-  }
-});
-
-Cypress.Commands.add('createAwxUser', (organization: Organization) => {
-  cy.requestPost<User>(`/api/v2/organizations/${organization.id.toString()}/users/`, {
-    username: 'e2e-user-' + randomString(4),
-    is_superuser: false,
-    is_system_auditor: false,
-    password: 'pw',
-    user_type: 'normal',
-  }).then((user) => user);
-});
-
-Cypress.Commands.add('deleteAwxUser', (user: User) => {
-  if (user.id) {
-    cy.requestDelete(`/api/v2/users/${user.id}/`, true);
-  }
-});
-
-Cypress.Commands.add('createAwxProject', () => {
-  cy.createAwxOrganization().then((organization) => {
-    cy.requestPost<Project>('/api/v2/projects/', {
-      name: 'E2E Project ' + randomString(4),
-      organization: organization.id,
-      scm_type: 'git',
-      scm_url: 'https://github.com/ansible/ansible-tower-samples',
-    }).then((project) => {
-      waitForProjectToFinishSyncing(project.id);
-    });
-  });
-});
-
-Cypress.Commands.add('deleteAwxProject', (project: Project) => {
-  const organizationId = project.organization;
-  // Delete sync job related to project
-  if (project && project.related && typeof project.related.last_job === 'string') {
-    const projectUpdateEndpoint: string = project.related.last_job;
-    cy.requestDelete(projectUpdateEndpoint);
-  }
-  // Delete project
-  cy.requestDelete(`/api/v2/projects/${project.id}/`, true);
-  // Delete organization for the project
-  cy.requestDelete(`/api/v2/organizations/${organizationId.toString()}/`, true);
-});
-
-Cypress.Commands.add('createAwxInventory', () => {
-  cy.createAwxOrganization().then((organization) => {
-    cy.requestPost<Inventory>('/api/v2/inventories/', {
-      name: 'E2E Inventory ' + randomString(4),
-      organization: organization.id,
-    }).then((inventory) => inventory);
-  });
-});
-
-Cypress.Commands.add('deleteAwxInventory', (inventory: Inventory) => {
-  const organizationId = inventory.organization;
-  // Delete organization created for this inventory (this will also delete the inventory)
-  if (organizationId) {
-    cy.requestDelete(`/api/v2/organizations/${organizationId.toString()}/`, true);
-  }
-});
-
-Cypress.Commands.add('createAwxJobTemplate', () => {
-  cy.createAwxProject().then((project) => {
-    cy.createAwxInventory().then((inventory) => {
-      cy.requestPost<JobTemplate>('/api/v2/job_templates/', {
-        name: 'E2E Job Template ' + randomString(4),
-        playbook: 'hello_world.yml',
-        project: project.id.toString(),
-        inventory: inventory.id,
-      }).then((jobTemplate) => jobTemplate);
-    });
-  });
-});
-
-Cypress.Commands.add('deleteAwxJobTemplate', (jobTemplate: JobTemplate) => {
-  const projectId = jobTemplate.project;
-
-  if (jobTemplate.id) {
-    const templateId = typeof jobTemplate.id === 'number' ? jobTemplate.id.toString() : '';
-    cy.requestDelete(`/api/v2/job_templates/${templateId}/`, true);
-  }
-  if (projectId) {
-    cy.requestGet<Project>(`/api/v2/projects/${projectId}/`).then((project) => {
-      // This will take care of deleting the project and the associated org, inventory
-      cy.deleteAwxProject(project);
-    });
-  }
-});
-
-let requestCount = 1;
-
-// Polling to wait till a project is synced
-function waitForProjectToFinishSyncing(projectId: number) {
-  cy.requestGet<Project>(`/api/v2/projects/${projectId}`).then((project) => {
-    // Assuming that projects could take up to 5 min to sync if the instance is under load with other jobs
-    if (project.status === 'successful' || requestCount > 300) {
-      if (requestCount > 300) {
-        cy.log('Reached maximum number of requests for reading project status');
-      }
-      // Reset request count
-      requestCount = 1;
-      return;
-    }
-    requestCount++;
-    cy.wait(1000);
-    waitForProjectToFinishSyncing(projectId);
-  });
-}
-
-Cypress.Commands.add(
-  'createInventoryHostGroup',
-  function createInventoryHostGroup(organization: Organization) {
-    cy.requestPost<Inventory>('/api/v2/inventories/', {
-      name: 'E2E Inventory ' + randomString(4),
-      organization: organization.id,
-    }).then((inventory) => {
-      cy.requestPost<Host>('/api/v2/hosts/', {
-        name: 'E2E Host ' + randomString(4),
-        inventory: inventory.id,
-      }).then((host) => {
-        cy.requestPost<{ name: string; inventory: number }>(`/api/v2/hosts/${host.id!}/groups/`, {
-          name: 'E2E Group ' + randomString(4),
-          inventory: inventory.id,
-        }).then((group) => ({
-          inventory,
-          host,
-          group,
-        }));
-      });
-    });
-  }
-);
-
-Cypress.Commands.add('createAwxLabel', (organization: Organization) => {
-  cy.requestPost<Label>('/api/v2/labels/', {
-    name: 'E2E Label ' + randomString(4),
-    organization: organization.id,
-  }).then((label) => label);
-});
-
-Cypress.Commands.add('deleteAwxLabel', (label: Label) => {
-  const labelId = label.id;
-  if (labelId) {
-    cy.requestDelete(`/api/v2/labels/${labelId.toString()}/`, true);
-  }
-});
-
-Cypress.Commands.add('createAwxInstanceGroup', () => {
-  cy.requestPost<InstanceGroup>('/api/v2/instance_groups/', {
-    name: 'E2E Instance Group ' + randomString(4),
-  }).then((instanceGroup) => instanceGroup);
-});
-
-Cypress.Commands.add('deleteAwxInstanceGroup', (instanceGroup: InstanceGroup) => {
-  const instanceGroupId = instanceGroup.id;
-  if (instanceGroupId) {
-    cy.requestDelete(`/api/v2/instance_groups/${instanceGroupId.toString()}/`, true);
-  }
-});
-
-/*  EDA related custom command implementation  */
-
-Cypress.Commands.add('edaRuleBookActivationActions', (action: string, rbaName: string) => {
-  cy.contains('td[data-label="Name"]', rbaName)
-    .parent()
-    .within(() => {
-      cy.get('button.toggle-kebab').click();
-      cy.contains('a', action).click();
-    });
-});
-
-Cypress.Commands.add('edaRuleBookActivationActionsModal', (action: string, rbaName: string) => {
-  cy.get('div[role="dialog"]').within(() => {
-    cy.contains('h1', `${action} activation`).should('be.visible');
-    cy.contains('p', `Are you sure you want to ${action} the rulebook activation below?`, {
-      matchCase: false,
-    }).should('be.visible');
-    cy.contains('p', rbaName).should('be.visible');
-    cy.contains('button#confirm', action).click();
-  });
-});
-
-Cypress.Commands.add('createEdaProject', () => {
-  cy.requestPost<EdaProject>('/api/eda/v1/projects/', {
-    name: 'E2E Project ' + randomString(4),
-    url: 'https://github.com/Alex-Izquierdo/eda-sample-project',
-  }).then((edaProject) => {
-    Cypress.log({
-      displayName: 'EDA PROJECT CREATION :',
-      message: [`Created 👉  ${edaProject.name}`],
-    });
-    return edaProject;
-  });
-});
-
-Cypress.Commands.add('getEdaRulebooks', (_edaProject) => {
-  // TODO: Once the API supports it, only get rulebooks for the project
-  // Sometimes it takes a while for the rulebooks to be created.
-  // Poll for EdaRulebooks until some are found.
-  cy.pollEdaResults<EdaRulebook>('/api/eda/v1/rulebooks/').then((edaRulebooks) => {
-    return edaRulebooks;
-  });
-});
-
-Cypress.Commands.add(
-  'createEdaRulebookActivation',
-  (edaRulebookActivation: SetOptional<Omit<EdaRulebookActivation, 'id'>, 'name'>) => {
-    cy.requestPost<EdaRulebookActivation>(`/api/eda/v1/activations/`, {
-      name: 'E2E Rulebook Activation ' + randomString(5),
-      restart_policy: 'on-failure',
-      ...edaRulebookActivation,
-    }).then((edaRulebookActivation) => {
-      cy.wrap(edaRulebookActivation)
-        .should('not.be.undefined')
-        .then(() => {
-          Cypress.log({
-            displayName: 'EDA RULEBOOK ACTIVATIONS CREATION :',
-            message: [`Created 👉  ${edaRulebookActivation.name}`],
-          });
-          return edaRulebookActivation;
-        });
-    });
-  }
-);
-
-Cypress.Commands.add('getEdaProject', (projectName: string) => {
-  cy.requestGet<EdaResult<EdaProject>>(`/api/eda/v1/projects/?name=${projectName}`).then(
-    (result) => {
-      if (result?.results && result.results.length === 1) {
-        return result.results[0];
-      }
-      return undefined;
-    }
-  );
-});
-
-Cypress.Commands.add('getEdaRulebookActivation', (edaRulebookActivationName: string) => {
-  cy.pollEdaResults<EdaRulebookActivation>(
-    `/api/eda/v1/activations/?name=${edaRulebookActivationName}`
-  ).then((activations) => {
-    return activations[0];
-  });
-});
-
-Cypress.Commands.add('deleteEdaRulebookActivation', (edaRulebookActivation) => {
-  cy.requestDelete(`/api/eda/v1/activations/${edaRulebookActivation.id}/`, true).then(() => {
-    Cypress.log({
-      displayName: 'EDA RULEBOOK ACTIVATION DELETION :',
-      message: [`Deleted 👉  ${edaRulebookActivation.name}`],
-    });
-  });
-});
-
-Cypress.Commands.add('waitEdaProjectSync', (edaProject) => {
-  cy.requestGet<EdaResult<EdaProject>>(`/api/eda/v1/projects/?name=${edaProject.name}`).then(
-    (result) => {
-      if (Array.isArray(result?.results) && result.results.length === 1) {
-        const project = result.results[0];
-        if (project.import_state !== 'completed') {
-          cy.wait(100).then(() => cy.waitEdaProjectSync(edaProject));
-        } else {
-          cy.wrap(project);
-        }
-      } else {
-        cy.wait(100).then(() => cy.waitEdaProjectSync(edaProject));
-      }
-    }
-  );
-});
-
-Cypress.Commands.add('getEdaProjectByName', (edaProjectName: string) => {
-  cy.requestGet<EdaResult<EdaProject>>(`/api/eda/v1/projects/?name=${edaProjectName}`).then(
-    (result) => {
-      if (Array.isArray(result?.results) && result.results.length === 1) {
-        return result.results[0];
-      } else {
-        return undefined;
-      }
-    }
-  );
-});
-
-Cypress.Commands.add('deleteEdaProject', (project: EdaProject) => {
-  cy.waitEdaProjectSync(project);
-  cy.requestDelete(`/api/eda/v1/projects/${project.id}/`, true).then(() => {
-    Cypress.log({
-      displayName: 'EDA PROJECT DELETION :',
-      message: [`Deleted 👉  ${project.name}`],
-    });
-  });
-});
-
-Cypress.Commands.add('pollEdaResults', (url: string) => {
-  cy.requestGet<EdaResult<unknown>>(url).then((result) => {
-    if (Array.isArray(result?.results) && result.results.length > 0) {
-      cy.wrap(result.results);
-    } else {
-      cy.wait(100).then(() => cy.pollEdaResults(url));
-    }
-  });
-});
-
-Cypress.Commands.add('createEdaCredential', () => {
-  cy.requestPost<EdaCredential>('/api/eda/v1/credentials/', {
-    name: 'E2E Credential ' + randomString(4),
-    credential_type: 'Container Registry',
-    description: 'This is a container registry credential',
-    username: 'admin',
-  }).then((edaCredential) => {
-    Cypress.log({
-      displayName: 'EDA CREDENTIAL CREATION :',
-      message: [`Created 👉  ${edaCredential.name}`],
-    });
-    return edaCredential;
-  });
-});
-
-Cypress.Commands.add('deleteEdaCredential', (credential: EdaCredential) => {
-  cy.requestDelete(`/api/eda/v1/credentials/${credential.id}/`, true).then(() => {
-    Cypress.log({
-      displayName: 'EDA CREDENTIAL DELETION :',
-      message: [`Deleted 👉  ${credential.name}`],
-    });
-  });
-});
-
-Cypress.Commands.add('getEdaCredentialByName', (edaCredentialName: string) => {
-  cy.requestGet<EdaResult<EdaCredential>>(
-    `/api/eda/v1/credentials/?name=${edaCredentialName}`
-  ).then((result) => {
-    if (Array.isArray(result?.results) && result.results.length === 1) {
-      return result.results[0];
-    } else {
-      return undefined;
-    }
-  });
-});
-
-Cypress.Commands.add('createEdaUser', () => {
-  cy.requestPost<EdaUser>(`/api/eda/v1/users/`, {
-    username: `E2E User ${randomString(4)}`,
-    email: `${randomString(4)}@redhat.com`,
-    password: `${randomString(4)}`,
-    type: 'super',
-  }).then((edaUser) => {
-    Cypress.log({
-      displayName: 'EDA USER CREATION :',
-      message: [`Created 👉  ${edaUser.username}`],
-    });
-    return edaUser;
-  });
-});
-
-Cypress.Commands.add('deleteEdaUser', (user: EdaUser) => {
-  cy.requestDelete(`/api/eda/v1/credentials/${user.id}/`, true).then(() => {
-    Cypress.log({
-      displayName: 'EDA CREDENTIAL DELETION :',
-      message: [`Deleted 👉  ${user.username}`],
-    });
-  });
-});
-
-Cypress.Commands.add('getEdaUser', () => {
-  cy.requestGet<EdaResult<EdaUser>>(`/api/eda/v1/users/me/`).then((result) => {
-    // This will take care of deleting the project and the associated org, inventory
-    if (Array.isArray(result?.results) && result?.results.length === 1) {
-      return result?.results[0];
-    } else {
-      return undefined;
-    }
-  });
-});
-Cypress.Commands.add('createEdaDecisionEnvironment', () => {
-  cy.requestPost<EdaDecisionEnvironment>('/api/eda/v1/decision-environments/', {
-    name: 'E2E Decision Environment ' + randomString(4),
-    image_url: 'quay.io/ansible/ansible-rulebook:main',
-  }).then((edaDE) => {
-    Cypress.log({
-      displayName: 'EDA DECISION CREATION :',
-      message: [`Created 👉  ${edaDE.name}`],
-    });
-    return edaDE;
-  });
-});
-
-Cypress.Commands.add('getEdaDecisionEnvironmentByName', (edaDEName: string) => {
-  cy.requestGet<EdaResult<EdaDecisionEnvironment>>(
-    `/api/eda/v1/decision-environments/?name=${edaDEName}`
-  ).then((result) => {
-    if (Array.isArray(result?.results) && result.results.length === 1) {
-      return result.results[0];
-    } else {
-      return undefined;
-    }
-  });
-});
-
-Cypress.Commands.add(
-  'deleteEdaDecisionEnvironment',
-  (decisionEnvironment: EdaDecisionEnvironment) => {
-    //cy.waitEdaDESync(decisionEnvironment);
-    cy.requestDelete(`/api/eda/v1/decision-environments/${decisionEnvironment.id}/`, true).then(
-      () => {
-        Cypress.log({
-          displayName: 'EDA DECISION ENVIRONMENT DELETION :',
-          message: [`Deleted 👉  ${decisionEnvironment.name}`],
-        });
-      }
-    );
-  }
-);
-
-/*
-Cypress.Commands.add('waitEdaDESync', (decisionEnvironment) => {
-  cy.requestGet<EdaResult<EdaDecisionEnvironment>>(
-    `/api/eda/v1/decision-environments/?name=${decisionEnvironment.name}`
-  ).then((result) => {
-    if (Array.isArray(result?.results) && result.results.length === 1) {
-      const project = result.results[0];
-      if (project.import_state !== 'completed') {
-        cy.wait(100).then(() => cy.waitEdaDESync(decisionEnvironment));
-      } else {
-        cy.wrap(project);
-      }
-    } else {
-      cy.wait(100).then(() => cy.waitEdaDESync(decisionEnvironment));
-    }
-  });
-});
-*/
